Print a message after child died to supported signal
Without `strsignal` or `psignal` we'd need a hardcoded table for all
possible signals. Just handle the ones we're most likely to encounter
and leave the user to figure out others with e.g. `echo $?`.

We could handle none of them, but then prompt ends up in the same line
as the echoed `^C` or `^\`. We could just print a newline if we have a
caught SIGINT before reading new input, but we really don't want to add
a SIGQUIT handler as then readline would echo '^\' whenever the keys are
pressed even though we do nothing with it.